backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
google-generativeai==0.5.4
python-dotenv==1.0.1
pydantic==2.7.1